Add new api for more than 64-bit hash
Add a hash function interface `LongTupleHashFunction` which extends the 64-bit `LongHashFunction`.
The new interface add two set of APIs while keeping compatible to the original long API:

1. long hash(..., long[] result);
2. long[] hashTuple(...);

Among sets of the hash API, the real work will appear in the `long hash(..., long[] result)' style
functions which should returns `result[0]' as the direct result.
The other sets can be deduced from this one.

In the mean time, the implementation of `LongTupleHashFunction` need override `bits()' function
providing the real bits of this function. The result of `bits()' should be larger than 64,
but not have to be divided by 64.

I choose `long[]' for the result container since it is very simple and can be optimised by JIT.
When developing the `long hash(..., long[] result)', we just need to ckeck if the result is null,
which means omitting the full result, and don't have to check the length of `result'.
`LongTupleHashFunction' provides a `newLongTuple()' function to alloc a correct array based on `bits()'.

Since a result container can be easily reused within one thread,
so the cost of allocation could trend to zero.

In this commit, the Murmur3F 128-bit has been migrated to `LongTupleHashFunction'.